+++ src/frontend/apps/standalone_site/src/features/Contents/features/Video/components/VideoRouter.tsx
@@ -5,10 +5,12 @@ import { routes } from 'routes';
 
 import VideoCreate from './Create/VideoCreate';
 import Videos from './Read/Videos';
+import VideoUpdate from './Update/VideoUpdate';
 
 const VideoRouter = () => {
   const videoRoute = routes.CONTENTS.subRoutes.VIDEO;
   const videoCreatePath = videoRoute.subRoutes?.CREATE?.path || '';
+  const videoUpdatePath = videoRoute.subRoutes?.UPDATE?.path || '';
 
   return (
     <Box pad="medium">
@@ -16,6 +18,9 @@ const VideoRouter = () => {
         <Route path={videoCreatePath} exact>
           <VideoCreate />
           <Videos />
+        </Route>
+        <Route path={videoUpdatePath} exact>
+          <VideoUpdate />
         </Route>
         <Route>
           <VideoCreate />